renamed mock database modules
