+++ tests/fixtures/databases/alexa/fixture_alexa_database.py
@@ -8,7 +8,7 @@ from AppLambda.src.clients.alexa import ListManagerClient
 from AppLambda.src.models.alexa import AlexaListItemOut, AlexaListOut, ListItemState, ListState
 from tests.utils import random_bool, random_int, random_string
 
-from .router import MockAlexaServer
+from .mock_alexa_database import MockAlexaServer
 
 _mock_alexa_server = MockAlexaServer()
